pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "bluebikes_analysis"
version = "0.0.1"
description = "Exploratory analysis of BlueBikes hitorical trip data, and demand forecasting development."
authors = [
  { name = "Juan Cruz Ferreyra" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "black",
    "flake8",
    "isort",
    "pip",
    "python-dotenv",
    "pandas (>=3.0.0,<4.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pytz (>=2025.2,<2026.0)",
]
requires-python = "~=3.11.0"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.isort]
profile = "black"
known_first_party = ["bluebikes_analysis"]
force_sort_within_sections = true


